Рабоатет все кроме схем проезда

+++ src/main/resources/liquibase/scripts/shelter_data_base.sql
@@ -3,7 +3,7 @@
 --changeset gturumtaev:1
 
 CREATE TABLE shelter (
-     id SERIAL PRIMARY KEY,
+     id INT8 PRIMARY KEY,
      work_schedule VARCHAR(255) NOT NULL,
      address VARCHAR(255) NOT NULL,
      phone_number_security VARCHAR(255) NOT NULL,
@@ -11,26 +11,26 @@ CREATE TABLE shelter (
 );
 
 CREATE TABLE driving_directions (
-     id SERIAL PRIMARY KEY,
+     id INT8 PRIMARY KEY,
      file_path VARCHAR(255) NOT NULL,
-     shelter_id INT,
+     shelter_id INT8,
      FOREIGN KEY (shelter_id) REFERENCES shelter(id)
 );
 
 --changeset gturumtaev:2
 
 CREATE TABLE volunteers (
-     id BIGINT PRIMARY KEY,
+     id INT8 PRIMARY KEY,
      first_name TEXT NOT NULL,
      phone_number TEXT NOT NULL,
-     shelter_id int,
+     shelter_id INT8,
      FOREIGN KEY (shelter_id) REFERENCES shelter(id)
 );
 
 --changeset gturumtaev:3
 
 CREATE TABLE clients (
-     id SERIAL PRIMARY KEY,
+     id INT8 PRIMARY KEY,
      first_name TEXT,
      phone_number TEXT,
      took_a_pet boolean,
@@ -41,7 +41,7 @@ CREATE TABLE clients (
 --changeset gturumtaev:4
 
 CREATE TABLE report_tg (
-     id SERIAL PRIMARY KEY,
+     id INT8 PRIMARY KEY,
      date_added timestamp NOT NULL,
      general_well_being TEXT,
      photo_name TEXT NOT NULL,
@@ -52,7 +52,7 @@ CREATE TABLE report_tg (
 --changeset gturumtaev:5
 
 CREATE TABLE animal_dog (
-     id SERIAL PRIMARY KEY,
+     id INT8 PRIMARY KEY,
      dog_name TEXT,
      age INT,
      color TEXT,
@@ -65,7 +65,7 @@ CREATE TABLE animal_dog (
 --changeset gturumtaev:6
 
 CREATE TABLE animal_cat (
-    id SERIAL PRIMARY KEY,
+    id INT8 PRIMARY KEY,
     cat_name TEXT,
     age INT,
     color TEXT,
@@ -78,7 +78,7 @@ CREATE TABLE animal_cat (
 --changeset gturumtaev:7
 
 CREATE TABLE owner (
-    id SERIAL PRIMARY KEY,
+    id INT8 PRIMARY KEY,
     client_name TEXT,
     animal_name TEXT,
     date_owner timestamp
